+++ CHANGELOG.md
@@ -34,6 +34,13 @@
 
 - Prepare projection query parameter converting it to json string [PR#101](https://github.com/quality-match/hari-client/pull/101)
 
+### Internal
+- Add internal methods for getting annotation run metrics [PR#108](https://github.com/quality-match/hari-client/pull/108)
+  - get_annotation_run_metrics
+  - get_annotation_run_node_metrics
+  - get_annotation_run_project_status
+  - get_annotation_run_project
+
 #### Consistency in partially failed uploads
 
 - partially failed bulk uploads of medias and media objects are now handled consistently [PR#80](https://github.com/quality-match/hari-client/pull/80)